Add files via upload
Summary:
I created this game of Crabs to practice coding in Java using Object Oriented Programming principles. The project consists of two classes: Die.java and Crabs.java. The Die.java class defines a dice and consists of methods for rolling the dice and obtaining the face Value. The Crabs.java class uses the Die.java class to create and use dice as well as define conditions of the game. Currently, the game is executed out of the main method within the Crabs.java file.

Ideas for Improvement:
If I wished to create additional games, I could create a client.java class with limited visibility that allows the user to pick from several games.